✨ feat: type 정리 #93
