+++ src/common/api/useMyTeamQuery.ts
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from "@tanstack/react-query";
-import { MyTeamType } from "@/common/types/type";
 import axiosInstance from "@/common/utils/axiosInstance";
+import { MyTeamType } from "@/common/types/type";
 
 async function fetchMyTeam() {
   const { data } = await axiosInstance.get("/myteam");
