+++ src/common/types/type.ts
@@ -1,98 +1,5 @@
 import { ReactNode } from "react";
 
-// Match 타입 정의
-export interface MatchType {
-  content: ContentProps[];
-  pageInfo: PageInfoProps;
-}
-
-export interface ContentProps {
-  gameId: number;
-  homeTeam: string;
-  awayTeam: string;
-  stadium: string;
-  gameStartTime: string;
-  sport: string;
-  timeOnSale: string;
-  timeOffSale: string;
-}
-
-export interface PageInfoProps {
-  page: number;
-  size: number;
-  totalElements: number;
-  totalPages: number;
-}
-
-export interface MatchData {
-  awayTeamName: string;
-  gameId: number;
-  gameStartTime: string;
-  homeTeamName: string;
-  latitude: number;
-  longitude: number;
-  stadiumName: string;
-  timeOnSale: string;
-  timeOffSale: string;
-}
-
-export type SportName = {
-  sport: "야구" | "축구" | "배구" | "농구";
-};
-export interface InformationCardProp {
-  gameId: number;
-  homeTeamName: string;
-  awayTeamName: string;
-  gameStartTime: string;
-  stadiumName: string;
-  reservationStatus?: boolean;
-  latitude: number;
-  longitude: number;
-}
-
-export type GameHistoryType = { reservationId: number; game: InformationCardProp }[];
-
-export interface Seat {
-  id: number;
-  isReserved: boolean;
-  isSelected: boolean;
-  price: number;
-}
-export interface seatFormData {
-  section: string;
-}
-
-export interface SectionProps {
-  seatPosition: string;
-  seatPrice: number;
-}
-export interface SeatsResponse {
-  homeTeamName: string;
-  awayTeamName: string;
-  stadiumName: string;
-  gameStartTime: string;
-  content: SectionProps[];
-  seat: Seat[];
-}
-
-export interface sectionOfSeats {
-  seatPosition: string;
-  seatPrice: number;
-  availableSeat: number;
-}
-
-export interface MatchDataProps {
-  awayTeamName: string;
-  gameId: number;
-  gameStartTime: string;
-  homeTeamName: string;
-  id: string;
-  sportName: string;
-  stadiumName: string;
-  timeOffSale: string;
-  timeOnSale: string;
-}
-
 export interface MyTeamType {
   id: string;
   teamName: string;
@@ -111,17 +18,4 @@ export interface LeftTime {
   hours: number;
   minutes: number;
   seconds: number;
-}
-export interface SeatType {
-  seatPosition: string;
-  seatRow: number;
-  seatNum: number;
-}
-export interface GameReservationType {
-  createdAt: string;
-  memberName: string;
-  reservationStatus: "CANCELED" | "COMPLETED";
-  seats: SeatType[];
-  totalPrice: number;
-  game: InformationCardProp;
 }
