+++ src/hooks/useInfiniteScroll.ts
@@ -1,4 +1,4 @@
-import { GameHistoryType, PageInfoProps } from "@/common/types/type";
+import { GameHistoryType, PageInfoProps } from "@/common/types/matchTypes";
 import alertToast from "@/common/utils/alertToast";
 import {
   FetchNextPageOptions,
