+++ src/pages/Home/api/useMainMatch.ts
@@ -1,12 +1,12 @@
+import { MainMatchType } from "@/common/types/matchTypes";
 import axiosInstance from "@/common/utils/axiosInstance";
-import { MainMatchType } from "@/pages/Home/components/MainMatch";
 import { useSuspenseQuery } from "@tanstack/react-query";
 
 export default function useMainMatch() {
   return useSuspenseQuery<MainMatchType>({
     queryKey: ["mainMatch"],
     queryFn: async () => {
-      return await axiosInstance.get("/games/mostPopular").then((res) => res.data);
+      return await axiosInstance.get("/games/mostPopularasd").then((res) => res.data);
     },
   });
 }
