+++ src/pages/Home/api/useWeeklyMatch.ts
@@ -1,4 +1,4 @@
-import { MatchData } from "@/common/types/type";
+import { MatchData } from "@/common/types/matchTypes";
 import axiosInstance from "@/common/utils/axiosInstance";
 import { useSuspenseQuery } from "@tanstack/react-query";
