+++ src/pages/Reservation/index.tsx
@@ -3,10 +3,9 @@ import { useLocation, useParams } from "react-router-dom";
 import { useForm, Controller } from "react-hook-form";
 import useAxios from "../../hooks/useAxios";
 import Section from "./components/seats/Section";
-import { Seat, seatFormData } from "../../common/types/type";
+import { seatFormData, Seat } from "@/common/types/seatTypes";
 import SectionOfSeats from "./components/seats/SectionOfSeats";
 import SelectedSeats from "./components/seats/SelectedSeat";
-
 import Loading from "@/common/components/atoms/Loading";
 import { seatDataProps, positionInfoProps } from "./type";
